refactor: move decorator to skip integration tests into test utils and extend status code parameter
